+++ src/JShark/Json.hs
@@ -2,6 +2,7 @@
     DataKinds
   , OverloadedStrings
 #-}
+
 -- | Minimal @JSON@ wrapper.
 module JShark.Json
   ( stringify