+++ src/JShark/Timers.hs
@@ -2,6 +2,7 @@
     DataKinds
   , OverloadedStrings
 #-}
+
 -- | Wrappers over the JS timer functions.
 module JShark.Timers
   ( setTimeout